Move session controller to api folder
